use a trait to avoid code duplication
